Commented more code, added check to see if path is file or folder
Assumed that path(s) passed were to files.  Now attempts to get the
file from path, if an exception is thrown assumes that the path is to a
folder (maybe use a better solution?  Seems like it'd be a little slow
going through to that parent folder to get a file, only to have it
throw an exception, then do it all again)

TODO: Add recursive folder copying